requirements.txt
Django==4.2.3
django-autoslug==1.9.9
django-tinymce==3.6.1